Add Regal for linting Rego
This PR adds [Regal](https://github.com/styraInc/regal) for linting the Rego
contained in this repo.

Practically, this PR adds a Regal configuration file adapted to the project,
where some of the most reported style issues are ignoed for the time being.
Additionally, we're adding a linter job to the CI pipeline to ensure future
updates to policy if compliant as well.

A few (hopefully) uncontroversial issues reported have also been fixed:

* [constant-condition](https://docs.styra.com/regal/rules/bugs/constant-condition)
* [use-assignment-operator](https://docs.styra.com/regal/rules/style/use-assignment-operator)
* [use-in-operator](https://docs.styra.com/regal/rules/idiomatic/use-in-operator)
* [use-some-for-output-vars](https://docs.styra.com/regal/rules/idiomatic/use-some-for-output-vars)
* [non-raw-regex-pattern](https://docs.styra.com/regal/rules/idiomatic/non-raw-regex-pattern)

While the changes are non-intrusive, and should have no impact on evaluation,
I naturally wanted to run the tests included in this repo and followed the instructions
in TESTING.md. However, even after having installed all of the dependencies listed as
required, the test command fails due to a missing `oc` command. I tried to find that
and possibly update the TESTING.md file to include it, but the instructions I followed
suggested logging in to a "RedHat Customer Portal" to get it. If being a customer is required
to run the tests, that requirement would be good to add to the docs as well.

Signed-off-by: Anders Eknert <anders@eknert.com>